fix: Move stacktrace trimming to trimmingprocessor
This is done such that stacktraces are trimmed after processing (most
importantly minidump processing)
